src/main/java/es/urjccode/mastercloudapps/adcs/draughts/controllers/Logic.java
package es.urjccode.mastercloudapps.adcs.draughts.controllers;

import java.util.HashMap;
import java.util.Map;

import es.urjccode.mastercloudapps.adcs.draughts.controllers.PlayController;
import es.urjccode.mastercloudapps.adcs.draughts.controllers.ResumeController;
import es.urjccode.mastercloudapps.adcs.draughts.controllers.StartController;
import es.urjccode.mastercloudapps.adcs.draughts.models.Game;
import es.urjccode.mastercloudapps.adcs.draughts.models.State;
import es.urjccode.mastercloudapps.adcs.draughts.models.StateValue;

public class Logic {

    private Game game;
	private State state;
	
	private Map<StateValue, AcceptorController> controllers;

	public Logic() {
		this.game = new Game();
		this.state = new State();
		this.controllers = new HashMap<>();
        this.controllers.put(StateValue.INITIAL, new StartController(this.game, this.state));
        this.controllers.put(StateValue.IN_GAME, new PlayController(this.game, this.state));
        this.controllers.put(StateValue.FINAL, new ResumeController(this.game, this.state));
        this.controllers.put(StateValue.EXIT, null);
    }
    
    public StateValue getStateValue() {
		return this.state.getValue();
	}
	
	public void next() {
		this.state.next();
	}

	public AcceptorController getController() {
        return this.controllers.get(this.state.getValue());
    }
	
}